% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abundance_func.R
\name{abundance_func}
\alias{abundance_func}
\title{Calculating abundances with adjusted survivals
This function uses matrix multiplication to calculate abundances from
transition matrices and N0 data}
\usage{
abundance_func(N0_rand, surv_trend, fert_mat)
}
\arguments{
\item{N0_rand}{a random matrix of 10 years of initial population data}

\item{surv_trend}{calculated from trend function}

\item{fert_mat}{fertility matrix from COMPADRE database}
}
\description{
Calculating abundances with adjusted survivals
This function uses matrix multiplication to calculate abundances from
transition matrices and N0 data
}
\author{
Elizabeth Hiroyasu
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).

Martha M. Ellis, Jennifer L. Williams, Peter Lesica, Timothy J. Bell, Paulette Bierzychudek, Marlin Bowles,
 Elizabeth E. Crone, Daniel F. Doak, Johan Ehrlen, Albertine Ellis-Adam, Kathryn McEachern, Rengaian Ganesan,
 Penelope Latham, Sheila Luijten, Thomas N. Kaye, Tiffany M. Knight, Eric S. Menges, William F. Morris,
 Hans den Nijs, Gerard Oostermeijer, Pedro F. Quintana-Ascencio, J. Stephen Shelly, Amanda Stanley, Andrea Thorpe,
 Tamara Ticktin, Teresa Valverde, and Carl Weekley. 2012. Matrix population models from 20 studies of perennial
 plant populations. Ecology 93:951.
}

