% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abundance_regr_func.R
\name{abundance_regr}
\alias{abundance_regr}
\title{abundance_regression}
\usage{
abundance_regr(abundance, years, active_stages)
}
\arguments{
\item{abundance}{array of abundance matrices calculated from transition matrices
with modified survival}

\item{years}{the number of years in the simulation}
}
\description{
Runs a regression analysis examining the power of detecting a trend in
modified abundance matrices
}
\author{
Elizabeth Hiroyasu
}

