% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_N0.R
\name{gen_N0}
\alias{gen_N0}
\title{gen_N0}
\usage{
gen_N0(N0, trans_mat)
}
\arguments{
\item{N0}{this is a vector of ones, and its dimensions are calculated from the dimensions of the transition matrix}

\item{trans_mat}{an array of transition matrices from the COMPADRE plant matrix database.}
}
\description{
Generating a matrix of N0 vectors to use in demographic trend simulations.
This function creates a matrix that is the average of given transition matrices from the COMPADRE plant matrix database.
It then #'calculates the population over a ten year period using this average transition matrix and an N0 vector of
ones.
}
\author{
Elizabeth Hiroyasu
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).
}

