% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Load_Compadre_Data.R
\name{Load_Compadre_Data}
\alias{Load_Compadre_Data}
\title{Read in a subset of the Compadre database}
\usage{
Load_Compadre_Data(CompadreFile)
}
\arguments{
\item{Compadre_file}{The name of the Compadre data file, in Rdata format. If the file is not in the working directory then this variable should also include the relative or absolute path.}
}
\value{
A list with metadata and mat slots following the same structure as the Compadre database but with only a subset of the data

 Notice that the first set of matrices (from Crone and Lesica 2004) appear to be
 unreliable, as the fecundity terms are all in the last column.
}
\description{
This function loads data from the compadre data set and subsets the data of interest
}
\author{
Elizabeth Hiroyasu
}
\references{
Code is adapted from Owen Jones & Rob Salguero-Gomez \url{https://github.com/jonesor/compadreDB/}

COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).
}

