% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Load_Data_func.R
\name{Load_Data}
\alias{Load_Data}
\title{Load_Data}
\usage{
Load_Data()
}
\description{
Loading compadre matrix data
This function loads data from the compadre data set
}
\author{
Elizabeth Hiroyasu
}
\references{
Code is adapted from Owen Jones & Rob Salguero-Gomez https://github.com/jonesor/compadreDB/

COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).
}

