% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Load_N0.R
\name{load_N0}
\alias{load_N0}
\title{load_N0
Loading N0 vectors from Ellis et al.}
\usage{
load_N0()
}
\description{
load_N0
Loading N0 vectors from Ellis et al.
}
\author{
Elizabeth Hiroyasu
}
\references{
Martha M. Ellis, Jennifer L. Williams, Peter Lesica, Timothy J. Bell, Paulette Bierzychudek, Marlin Bowles,
 Elizabeth E. Crone, Daniel F. Doak, Johan Ehrl?n, Albertine Ellis-Adam, Kathryn McEachern, Rengaian Ganesan,
 Penelope Latham, Sheila Luijten, Thomas N. Kaye, Tiffany M. Knight, Eric S. Menges, William F. Morris,
 Hans den Nijs, Gerard Oostermeijer, Pedro F. Quintana-Ascencio, J. Stephen Shelly, Amanda Stanley, Andrea Thorpe,
 Tamara Ticktin, Teresa Valverde, and Carl Weekley. 2012. Matrix population models from 20 studies of perennial
 plant populations. Ecology 93:951.

code is adapted from Ellis et al. (2012) metadata
}

