% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surv_regr.R
\name{surv_regr}
\alias{surv_regr}
\title{surv_regression}
\usage{
surv_regr(surv_trend)
}
\arguments{
\item{surv_trend}{array of survival matrices with inserted trend beta}
}
\description{
Runs a regression analysis examining the power of detecting a trend in
modified survival matrices
}
\author{
Elizabeth Hiroyasu
}

