% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trend_func.R
\name{trend}
\alias{trend}
\title{trend_func}
\usage{
trend(surv_rand, beta, nstage, stage2mod)
}
\arguments{
\item{surv_rand}{A randomly sampled array of survival matrices}

\item{beta}{dictates how large the trend is to be introduced to survival matrices}

\item{nstage}{is the number of stages represented in the matrix data}
}
\description{
Adding a trend to survival rates in different stages
}
\details{
This function adds a trend to stage 2 survival
in the survival matrices for a stage or age structured population.
}
\author{
Elizabeth Hiroyasu
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).
}

