% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_N0.R
\name{gen_N0}
\alias{gen_N0}
\title{Generate a matrix of N0 vectors to use in demographic trend simulations.}
\usage{
gen_N0(trans_mat, n_return = 1, n_converge = 100)
}
\arguments{
\item{trans_mat}{an array of transition matrices}

\item{n_return}{number of abundance vectors to be returned}

\item{n_converge}{number of time steps to simulate before recording abundance}
}
\value{
A matrix of population vectors. Even if \code{n_return = 1}, a
 1xnstage matrix is returned, for compatibilty with legacy code elsewhere (I
 hope this can be updated)
}
\description{
Generate one or more abundance vectors from the stochastic stable
distribution of a collection of matrices.
}
\details{
Starts the population at the stable stage distribution of the mean matrix. Then the population is iterated for \code{n_converge} years, drawing a matrix at random each year. Finally, the iteration is continued for \code{n_return} years, again with a random draw each year.

The default value of \code{n_converge} is currently arbitrary, and may be too small to guarantee convergence to the stochastic stable distribution

If \code{n_return > 1}, the vqlues will \emph{not} be indepdendent, as they are subsequent values from a single stochastic simulation.
}
\author{
Elizabeth Hiroyasu and Bruce Kendall
}
\references{
COMPADRE Plant Matrix Database. Max Planck Institute for
Demographic Research (Germany). Available at www.compadre-db.org
(data downloaded on [1/1/2014]).
}

